pyproject.toml
[tool.poetry]
name = "mindforge"
version = "1.0.0"
description = "Платформа для онлайн-обучения на Django REST Framework."
authors = ["Ilya Yaushkin <mexic2284@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
pillow = "^11.1.0"
python-dotenv = "^1.0.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
isort = "^6.0.0"
mypy = "^1.15.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
python_version = "3.12"

[tool.black]
# Максимальная длина строки
line-length = 119
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119